[data ingestion] add support for bigtable in main workflow binary
