+++ crates/sui-data-ingestion/Cargo.toml
@@ -31,6 +31,7 @@ tracing.workspace = true
 sui-archival.workspace = true
 sui-storage.workspace = true
 sui-data-ingestion-core.workspace = true
+sui-kvstore.workspace = true
 sui-types.workspace = true
 tempfile.workspace = true
 url.workspace = true
