+++ crates/sui-kvstore/src/lib.rs
@@ -4,6 +4,7 @@ mod bigtable;
 use anyhow::Result;
 use async_trait::async_trait;
 pub use bigtable::client::BigTableClient;
+pub use bigtable::worker::KvWorker;
 use sui_types::crypto::AuthorityStrongQuorumSignInfo;
 use sui_types::digests::TransactionDigest;
 use sui_types::effects::{TransactionEffects, TransactionEvents};
